updated the code remove some unwanted lines.

+++ applicants-child-post-type.php
@@ -3,7 +3,7 @@
 function child_post_type_applicants()
 {
     $supports = array(
-        //   'title',
+        // 'title',
         // 'editor',
         // 'author',
         // 'thumbnail',
@@ -33,9 +33,6 @@ function child_post_type_applicants()
         'has_archive' => true,
         'hierarchical' => false,
         'show_in_menu' => 'edit.php?post_type=jobs',
-        // 'capabilities' => array(
-        //     'create_posts' => false, // Removes support for the "Add New" function ( use 'do_not_allow' instead of false for multisite set ups )
-        // ),
     );
     register_post_type('applicants', $args);
 }
@@ -55,8 +52,6 @@ function submit_job_application()
         wp_die();
     }
 
-    // here store the submited data as post as each posts for the this post type
-    // Prepare post data
     $post_data = array(
         'post_title' => $applicant_Name,
         'post_content' => $message,
@@ -69,16 +64,10 @@ function submit_job_application()
         'child_post_type' => 'applicants',
         'parent_post_type' => 'jobs',
         'job_id' => $jobId,
-        // You can add any other custom fields here
     );
 
-    // Insert the post into the database
     $post_id = wp_insert_post($post_data);
-    write_log("Create new application: " . $post_id);
-
-    // Handle post insertion result
     if ($post_id) {
-        // Add meta data for applicant's email and job ID
         update_post_meta($post_id, 'job_id', $jobId);
         update_post_meta($post_id, 'applicant_name', $applicant_Name);
         update_post_meta($post_id, 'applicant_email', $applicant_Email);
@@ -88,30 +77,18 @@ function submit_job_application()
         update_post_meta($post_id, 'post_type', 'jobs');
         update_post_meta($post_id, 'application_id', $post_id);
 
-
-
         // store applicant data
         global $wpdb;
         $table_name = $wpdb->prefix . 'jm_applicants';
-        $date = date('Y-m-d H:i:s');
+        $date = current_time('mysql');
         $job_name = get_the_title($jobId);
 
         $query_result = $wpdb->query(
             $wpdb->prepare(
                 "
         INSERT INTO $table_name (`post_id`,`date`, `name`, `email`, `cover`, `job_id`, `job_name`, `status`)
-        VALUES (%s, %s, %s, %s, %s, %d, %s, %s)",
-                $post_id,
-                $date,
-                $applicant_Name,
-                $applicant_Email,
-                $message,
-                $jobId,
-                $job_name,
-                'pending'
-            )
-        );
-
+        VALUES ('$post_id', '$date', '$applicant_Name', '$applicant_Email', '$message','$jobId','$job_name', 'pending')"
+        ));
 
         if ($query_result !== false) {
             echo json_encode(
@@ -125,11 +102,9 @@ function submit_job_application()
         } else {
             echo json_encode(array('status' => 'error', 'message' => 'Failed to submit application. Error: ' . $wpdb->last_error));
         }
-
     } else {
         echo json_encode(array('status' => 'error', 'message' => 'Failed to submit application'));
     }
-
     wp_die();
 }
 add_action('wp_ajax_submit_job_application', 'submit_job_application');
@@ -144,16 +119,13 @@ function filter_cpt_columns($columns)
     $columns['email'] = __('Email');
     $columns['job_name'] = __('Applied Job');
     $columns['application_status'] = __('Application status');
-
     return $columns;
 }
 
-
 add_action('manage_posts_custom_column', 'action_custom_columns_content', 10, 2);
-function action_custom_columns_content($column_id, $post_id)
+function action_custom_columns_content($column_name, $post_id)
 {
-    //run a switch statement for all of the custom columns created
-    switch ($column_id) {
+    switch ($column_name) {
         case 'email':
             echo ($value = get_post_meta($post_id, 'applicant_email', true)) ? $value : '';
             break;
@@ -168,33 +140,16 @@ function action_custom_columns_content($column_id, $post_id)
     }
 }
 
-// hide add new applicants
-// function hd_add_buttons()
-// {
-//     global $pagenow;
-//     if (is_admin()) {
-//         if ($_GET['post_type'] == 'applicants') {
-//             echo '<style>.page-title-action{display: none !important;}</style>';
-//         }
-//     }
-// }
-// add_action('admin_head', 'hd_add_buttons');
-
-
 add_action('updated_post_meta', 'update_application');
-
-function update_application($post_id)
+function update_application()
 {
-    write_log("post ID: " . $post_id);
-
-    // $newStatus = $_POST['meta'];
+    // Verify nonce
+    if (!isset($_POST['_wpnonce'])) {
+        return;
+    }
 
     if (isset($_POST['meta'])) {
-        // 'meta' key exists in $_POST
         $meta_data = $_POST['meta'];
-
-
-        // Define variables to store values
         $applicant_email = '';
         $applicant_name = '';
         $application_cover = '';
@@ -205,28 +160,32 @@ function update_application($post_id)
         foreach ($meta_data as $meta_item) {
             switch ($meta_item['key']) {
                 case 'applicant_email':
-                    $applicant_email = $meta_item['value'];
+                    $applicant_email = sanitize_email($meta_item['value']);
                     break;
                 case 'applicant_name':
-                    $applicant_name = $meta_item['value'];
+                    $applicant_name = sanitize_text_field($meta_item['value']);
                     break;
                 case 'application_cover':
-                    $application_cover = $meta_item['value'];
+                    $application_cover = sanitize_textarea_field($meta_item['value']);
                     break;
                 case 'application_status':
-                    $application_status = $meta_item['value'];
+                    $application_status = sanitize_text_field($meta_item['value']);
                     break;
                 case 'application_id':
-                    $application_id = $meta_item['value'];
+                    $application_id = intval($meta_item['value']);
                     break;
             }
         }
+        
         // store applicant data
         global $wpdb;
         $table_name = $wpdb->prefix . 'jm_applicants';
         $update_query = "UPDATE `$table_name` SET `email`='$applicant_email', `name`= '$applicant_name', `cover`='$application_cover', `status`='$application_status' WHERE `post_id` = '$application_id'";
-        $wpdb->query($update_query);
-
+        $result = $wpdb->query($update_query);
+        if ($result === false) {
+            // Handle database error
+            error_log("Error occurred while updating applicant data with post ID: $application_id");
+        }
     } else {
         // 'meta' key does not exist in $_POST
         echo "No meta data found in the POST request.";
@@ -234,15 +193,17 @@ function update_application($post_id)
 }
 
 add_action('wp_trash_post', 'application_delete');
-
-function application_delete()
+function application_delete($post_id)
 {
-    if (isset($_GET['post'])) {
+    if (isset($_GET['post']) && wp_verify_nonce($_REQUEST['_wpnonce'], 'trash-post_' . $post_id)) {
         $post_id = intval($_GET['post']);
-        write_log("Delete Application ID: " . $post_id);
         global $wpdb;
         $table_name = $wpdb->prefix . 'jm_applicants';
         $update_query = "DELETE FROM $table_name WHERE `post_id` = '$post_id'";
-        $wpdb->query($update_query);
+        $result = $wpdb->query($update_query);
+        if ($result === false) {
+            // Handle the error
+            error_log("Error occurred while deleting entry with post ID: $post_id");
+        }
     }
 }
